// src/models/Plan/PlanItems.ts
export const PlanItem = t.model('PlanItemInn', {
  id: t.maybeNull(t.number),
  b_article_lvl3_color_row: t.maybeNull(t.number),
  b_correction_info: t.maybeNull(t.number),
  b_rw_ca: t.maybeNull(t.number),
  b_rw_tb: t.maybeNull(t.number),
  branch_code: t.maybeNull(t.string),
  branch_name: t.maybeNull(t.string),
  delta_1q: t.maybeNull(t.number),
  delta_2q: t.maybeNull(t.number),
  delta_3q: t.maybeNull(t.number),
  delta_4q: t.maybeNull(t.number),
  delta_9m: t.maybeNull(t.number),
  delta_y: t.maybeNull(t.number),
  fact_1q: t.maybeNull(t.number),
  fact_2m2q: t.maybeNull(t.number),
  fact_2q: t.maybeNull(t.number),
  fact_3q: t.maybeNull(t.number),
  fact_4q: t.maybeNull(t.number),
  fact_prev_1q: t.maybeNull(t.number),
  fact_prev_2q: t.maybeNull(t.number),
  fact_prev_3q: t.maybeNull(t.number),
  fact_prev_4q: t.maybeNull(t.number),
  fact_prev_9m: t.maybeNull(t.number),
  fact_prev_y: t.maybeNull(t.number),
  forecast_1q: t.maybeNull(t.number),
  forecast_2q: t.maybeNull(t.number),
  forecast_3q: t.maybeNull(t.number),
  forecast_4q: t.maybeNull(t.number),
  forecast_y: t.maybeNull(t.number),
  gkm_name: t.maybeNull(t.string),
  i_tb_name_min: t.maybeNull(t.string),
  id_article_lvl1: t.maybeNull(t.number),
  id_article_lvl2: t.maybeNull(t.number),
  id_article_lvl3: t.maybeNull(t.number),
  id_branch: t.maybeNull(t.number),
  id_currency: t.maybeNull(t.number),
  id_org: t.maybeNull(t.number),
  id_org_code: t.maybeNull(t.string),
  id_org_name: t.maybeNull(t.string),
  id_stage: t.maybeNull(t.number),
  id_tb: t.maybeNull(t.number),
  id_tb_inn: t.maybeNull(t.number),
  id_xo: t.maybeNull(t.number),
  n_article_lvl1_color: t.maybeNull(t.number),
  n_article_lvl1_order: t.maybeNull(t.number),
  n_article_lvl2_color: t.maybeNull(t.number),
  n_article_lvl2_order: t.maybeNull(t.number),
  n_article_lvl3_color: t.maybeNull(t.number),
  n_article_lvl3_order: t.maybeNull(t.number),
  plan_approved_1q: t.maybeNull(t.number),
  plan_approved_2q: t.maybeNull(t.number),
  plan_approved_3q: t.maybeNull(t.number),
  plan_approved_4q: t.maybeNull(t.number),
  plan_approved_9m: t.maybeNull(t.number),
  plan_approved_y: t.maybeNull(t.number),
  plan_ca_1q: t.maybeNull(t.number),
  plan_ca_2q: t.maybeNull(t.number),
  plan_ca_3q: t.maybeNull(t.number),
  plan_ca_4q: t.maybeNull(t.number),
  plan_ca_9m: t.maybeNull(t.number),
  plan_ca_y: t.maybeNull(t.number),
  plan_tb_1q: t.maybeNull(t.number),
  plan_tb_2q: t.maybeNull(t.number),
  plan_tb_3q: t.maybeNull(t.number),
  plan_tb_4q: t.maybeNull(t.number),
  plan_tb_9m: t.maybeNull(t.number),
  plan_tb_y: t.maybeNull(t.number),
  plan_xo_1q: t.maybeNull(t.number),
  plan_xo_2q: t.maybeNull(t.number),
  plan_xo_3q: t.maybeNull(t.number),
  plan_xo_4q: t.maybeNull(t.number),
  plan_xo_9m: t.maybeNull(t.number),
  plan_xo_y: t.maybeNull(t.number),
  s_article_lvl1_name: t.maybeNull(t.string),
  s_article_lvl2_name: t.maybeNull(t.string),
  s_article_lvl3_name: t.maybeNull(t.string),
  s_article_lvl3_msm: t.maybeNull(t.string),
  s_currency_code: t.maybeNull(t.string),
  xo_code: t.maybeNull(t.string),
  xo_name: t.maybeNull(t.string),
  plan_1q: t.maybeNull(t.number),
  plan_2q: t.maybeNull(t.number),
  plan_3q: t.maybeNull(t.number),
  plan_4q: t.maybeNull(t.number),
  plan_9m: t.maybeNull(t.number),
  plan_inn_1q: t.maybeNull(t.number),
  plan_inn_2q: t.maybeNull(t.number),
  plan_inn_3q: t.maybeNull(t.number),
  plan_inn_4q: t.maybeNull(t.number),
  plan_inn_9m: t.maybeNull(t.number),
  plan_inn_y: t.maybeNull(t.number),
  plan_y: t.maybeNull(t.number),
  s_branch_name: t.maybeNull(t.string),
});

// export type PlanItemInnInstance = Instance<typeof PlanItemInn>;
// export type PlanArticleItemInstance = Instance<typeof PlanArticleItem>;
// export type PlanItemInn = Instance<typeof PlanItemInn>

export type PlanItemInstance = Instance<typeof PlanItem>;